# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechNest Shop API
  version: 1.0.0
  description: >
    Versioned OpenAPI contract for the current TechNest Shop backend. This file is
    the authoritative machine-readable API contract; README keeps the operational
    snapshot and docs/API_CONTRACT.md keeps a concise human summary.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin Users]
      operationId: listUsers
      summary: List users
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin Users]
      operationId: createUser
      summary: Create user
      description: Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Admin Users]
      operationId: getUser
      summary: Get user by id
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Admin Users]
      operationId: updateUser
      summary: Update user
      description: Admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Admin Users]
      operationId: deleteUser
      summary: Delete user
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: User deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    BadRequest:
      description: Request validation failed or payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, status, error]
...
    UserRole:
      type: string
      enum: [ADMIN, STAFF, CUSTOMER, admin, staff, customer]
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        username:
          type: string
        fullName:
          type: string
        phone:
          type: string
        addressText:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
      required: [id, email, username, role]
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        username:
          type: string
          pattern: '^\\s*$|^[A-Za-z0-9._-]{3,50}$'
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
        fullName:
          type: string
          maxLength: 120
        phone:
          type: string
          pattern: '^\\s*$|^[0-9+()\\- ]{8,20}$'
        addressText:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
      required: [email, password, role]
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        username:
          type: string
          pattern: '^\\s*$|^[A-Za-z0-9._-]{3,50}$'
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
        fullName:
          type: string
          maxLength: 120
        phone:
          type: string
          pattern: '^\\s*$|^[0-9+()\\- ]{8,20}$'
        addressText:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
...